% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWRFC_clus.R
\name{GWRFC_clus}
\alias{GWRFC_clus}
\title{Clustering and statistical summaries}
\usage{
GWRFC_clus(
  input_shapefile,
  remove_columns = NA,
  clus_data = "all",
  method_hc = "ward.D2",
  clus_num = "auto",
  plots = T,
  scaled = T,
  output_folder
)
}
\arguments{
\item{input_shapefile}{string or Spatial-class. It can be a filename or an object of class {SpatialPolygonsDataFrame} or {SpatialPointsDataFrame}.}

\item{remove_columns}{string. Remove specific variables from \strong{input_shapefile}. Variables are identified by column name. NA ignores column remove.}

\item{clus_data}{string. Refers to data which should be used included for clustering. It can be "all" to use all \strong{input_shapefile} variables or any other variables specified. In all cases, it uses numeric variables except with "lucca" clustering approach. If one variable is declared, clustering is not applied and data from target variable is reported according its quantiles (when it is numerical) or its classes (when it is categorical).}

\item{method_hc}{string. A method to use for clustering. As hierchachical clustering is applied through {hclust} function. It can be:"ward.D","ward.D2","single","complete","average","mcquitty","median", "centroid" but other alternatives are "SOM" and "lucca" which uses kohonen and mclust packages respectively. Not relevant if \strong{clus_data} includes one variable.}

\item{clus_num}{numeric or string. If clustering is applied, this is the number of clusters. If it can be defined as "auto" (using Calinski-Harabasz Index) or any number greater than 2.}

\item{plots}{logical. If true, plots are generated and stored in the \strong{output_folder}.}

\item{scaled}{logical. If true, quantitative variables plots are scaled and showed as boxplots, otherwise as bars.}

\item{output_folder}{string. Output folder where results will be stored.}
}
\value{
The results of the function includes: \enumerate{
                           \item A shapefile with the resulting clusters or classes used for report.
                           \item A statistical summary for each cluster or class (stored in a .rds file)
                           \item Plots derived from the cluster analysis.
                          }
}
\description{
This function clusters table columns of a SpatialDataFrame object and reports its resulting clusters as an statistical summary. It is also capable to report a specific column but clustering is not applied and values are reported as quantiles (if numerical), or classes (if categorical).
}
\examples{
#based in the example showed with the execution of GWRFC
list.files("E:/demo/test",full.names=T,pattern=".shp") #get the output filename from GWRFC function

GWRFC_clus(input_shapefile = "E:/demo/test/GWRFC_ADP_400_EX_LVI.shp" , # we choose LVI output
         remove_columns = "ID_row",
         clus_data = "all" #to use all columns.
         method_hc="ward.D2",
         clus_num = "auto",
         plots=T,
         output_folder = "E:/DATA/demo/clus") #check this folder for outputs generated by the function.
}
