% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWRFC_repo.R
\name{GWRFC_repo}
\alias{GWRFC_repo}
\title{GWRFC outputs report}
\usage{
GWRFC_repo(
  dependent_varName,
  GWRFC_folder,
  plots = T,
  size_plots = c(22, 22),
  output_folder
)
}
\arguments{
\item{dependent_varName}{string. Dependent variable name. Only required for edit titles of plots. Recommended full names not longer than 8 words.}

\item{GWRFC_folder}{string. A folder where GWRFC outputs are stored. The function do not works if it finds more than an output.}

\item{plots}{logical. If true, plots are returned.}

\item{size_plots}{numeric. Base dimensions for plots. Set X and Y as centimeters: c(22,22)}

\item{output_folder}{string. Output folder where outputs will be stored.}
}
\value{
The results of the function includes: \enumerate{
                           \item A report summarizing GWRFC outputs (stored in a .rds file)
                           \item Plots derived from report summaries.
                          }
}
\description{
This function uses outputs from GWRFC and generate a report of dependent variable, using its classes to generate a statistical summary and plots.
}
\examples{
#based in the example showed with the execution of GWRFC
list.files("C:/DATA/demo/deforestation",full.names=T,pattern=".shp") #get the output filename from GWRFC function

GWRFC_repo(dependent_varName = "FAO deforestation rate (2001-2010)",
         GWRFC_folder = "E:/demo/test",
         plots = T,
         output_folder = "E:/demo/test/repo") #check this folder for outputs generated by the function.
}
