% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotJPEG.R
\name{plotJPEG}
\alias{plotJPEG}
\title{Plot JPEG file using base graphics}
\usage{
plotJPEG(input_jpeg, add = FALSE)
}
\arguments{
\item{input_jpeg}{string. Input filename of JPG picture.}
}
\description{
Internal function to plot a JPEG file in R
}
