% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWRFC.R
\name{GWRFC}
\alias{GWRFC}
\title{Geographically weighted Random Forest Classification (GWRFC)}
\usage{
GWRFC(input_shapefile, remove_columns = NA, dependent_varName,
  kernel_type = "exponential", kernel_adaptative = T, kernel_bandwidth,
  clusters_LVI = "auto", number_cores = 1, output_folder)
}
\arguments{
\item{input_shapefile}{spatialpolygonsdataframe. Input shapefile with dependent and independent variables}

\item{remove_columns}{string. Remove specific variables from 'input_shapefile'. Variables are identified by column name. NA ignores column remove.}

\item{dependent_varName}{string. Dependent variable name. Must exists at 'input_shapefile'.}

\item{kernel_type}{string. Kernel type to apply in GWRFC. It can be: 'gaussian', 'exponential', 'bisquare' or 'tricube'.}

\item{kernel_adaptative}{logical. Is the kernel adaptative? otherwise it is considered as fixed.}

\item{kernel_bandwidth}{numeric. Defines kernel bandwidth. If 'kernel_adaptative' is TRUE, define the number of local observations in the kernel, otherwise define its distance.}

\item{clusters_LVI}{numeric. Number of clusters for summarize local variables importance (LVI). If it is defined as 'auto' (default), it is calculated via Nbclust package.}

\item{number_cores}{numeric. Number of cores for parallel processing. Cores are register and operated via doParallel, foreach and parallel packages.}

\item{output_folder}{string. Output folder where GWRFC results are stored.}
}
\description{
CHECK...
}
