% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVIclust.R
\name{LVIclust}
\alias{LVIclust}
\title{Local variables importance (LVI) clustering from GWRFC outputs}
\usage{
LVIclust(input_shapefile, input_GWRFC, method_hc = "ward.D2",
  clus_data = "LVI", clus_num = "auto", plots = T, output_folder)
}
\arguments{
\item{input_shapefile}{string or Spatial-class. Input shapefile with dependent and independent variables. It can be the filename of the shapefile or an object of class SpatialPolygonsDataFrame or SpatialPointsDataFrame.}

\item{input_GWRFC}{string or Spatial-class. Input shapefile of GWRFC outputs. It can be the filename of the shapefile or an object of class SpatialPolygonsDataFrame or SpatialPointsDataFrame.}

\item{method_hc}{string. A method to use for hierarchical clustering with hclust. It can be:"ward.D","ward.D2","single","complete","average","mcquitty","median" or "centroid".}

\item{clus_data}{string. Data which should be used in clustering. It can be "LVI" to refer to all LVI variables or any specific column at input_GWRFC. In this case, hierarchical clustering is not applied and target variable is reclassified into quantiles if it is numerical, otherwise variables classes is used.}

\item{clus_num}{numeric or string. If applies hierarchical clustering, then it is the number of clusters; otherwise it is the number of quantiles to use. If it is defined as 'auto' (default), then a number is calculated via Calinski-Harabasz Index in hierarchical clustering or defined in 5 for quantiles.}

\item{plots}{logical. If true, plots are generated and stored in the output_folder.}

\item{output_folder}{string. Output folder where LVIclust outputs will be stored.}
}
\value{
The output of this function is a shapefile with the resulting clusters, a summary for each cluster stored in a .rds file, and optionally a series of plots to visualize the report.
}
\description{
This function summarize results from the GWRFC algorithm, clustering its main outputs (i.e. LVI) and using these results to report a summary based on its data source (i.e. the input_shapefile). It is also possible to use any other specific variable (e.g. probabilities of a target class) to generate reports but in ths case clustering is not applied and the variable is subset in quantiles (if numerical) or by classes (if categorical) to generate the report.
}
\examples{
#based in the example showed with the execution of GWRFC
list.files("C:/DATA/demo/",full.names=T,pattern=".shp") #get the output filename from GWRFC function

LVIclust(input_shapefile = deforestation, #should be the same shapefile used to apply GWRFC (do not need to remove variables)
         input_GWRFC = "C:/DATA/demo/GWRFC_ADP_300_exponential.shp", #filename of the GWRFC output
         method_hc="ward.D2", #hierarchical clustering is applied but see: help(hclust) for other methods.
         clus_data="LVI", #here is clustered all LVI results (34 in total, i.e. the number of variables used in the analysis).
         clus_num = 2, #number of clusters. It can be automatically calculated if it is 'auto'
         plots=T, #three plots referring to: LVI ranking by clusters, report for categorial and quantitative variables by cluster.
         output_folder = "C:/DATA/demo/") #check this folder for outputs generated by the function.
}
