% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVIclust.R
\name{LVIclust}
\alias{LVIclust}
\title{Local variables importance (LVI) clustering from GWRFC outputs}
\usage{
LVIclust(input_shapefile, input_GWRFC, method_hc = "ward.D2",
  num_clusters = "auto", plots = T, output_folder)
}
\arguments{
\item{input_shapefile}{string. Input shapefile with dependent and independent variables. It can be polygons or points based.}

\item{input_GWRFC}{string. Input shapefile of GWRFC outputs.}

\item{method_hc}{string. A method to uso for hierarchical clustering with hclust. It can be: "ward.D", "ward.D2","single","complete","average","mcquitty","median","centroid"}

\item{num_clusters}{numeric. Number of clusters for summarize local variables importance (LVI). If it is defined as 'auto' (default), it is calculated via Calinski-Harabasz Index and the second peak from 2,3,4...20 cluster computations. If numeric, it cannot be less than 2.}

\item{plots}{logical. If true, plots and summary reports for each cluster are created.}

\item{output_folder}{string. Output folder where GWRFC outputs will be stored.}
}
\description{
CHECK...
}
